[REF-IMP] survey : validate questions in frontend before submit
This commit adds the validation of all types of questions directly in the
frontend. This allows to reduce the latency before the survey is validated.
The client does not need to wait the validation answer from the server to
display validation error messages.

The current questions are only faded out if the frontend validation has passed.
But the server still makes this validation before submitting answers and
loading next question(s).

This commit also adds a slide down transition on error messages,
for a wonderfull wow effect !

Task ID: 2152223
PR #41453