[REF] survey : integrate breadcrumb in one page survey - ajax transition
As the survey form now works in ajax to display in same page start screen,
question(s) and end page screen, the breadcrumb has to be controlled not only
on page loading but also manually between transition from a screen to another.

To ease the breadcrumb management, breadcrumb is now a widget that will reload
a js template, mainly depending of the current page id.
The js form initialise or update this widget giving this new page id.
On breadcrumb item click, the widget triggers a onclick catched by the form
that calls the submit with the given target.

This commit handles page transition for the breadcrumb. When going forward or
backward (next or previous page), the breadcrumb active, inactives and links
elements are reinitialised depending of the target page.

Task ID: 2152223
PR #41453