[IMP] sale_stock: split the widget "qty_at_date_widget"
Purpose
=======
Use an abstract class for the popover widget, so we can create later others similar widgets.

How
===
The child widget inherit from "PopoverAbstract", and can add some option
- icon: icon displayed in the button
- title: title of the popover
- trigger: how the popover is triggered
- color: class used as color for the button
- popoverTemplate: name of the template used for the popover content
- hide: if true, the button is invisible
- placement: where the popover will spawn

Trick
=====
The method ``_willRender`` is call before rendering the widget.
If you want to implement some logic behind the color, icon, etc...
It's the method you search.
e.g.: In ``qty_at_date_widget``, the color depends on the ``virtual_available_at_date``.

We cannot do that in the ``init`` method, because in the ``one2many`` widget, the widget
is created, and then we can change the record...

Task #2153126