[IMP] mail: add author field on mail template
When you configure a mail template, you can specify an author for
the mail. This allows to simplify the setting of a template because
you no longer need to now the format to select the email of the author.
Example: for an invoice, you can set the author to "${user.partner_id.id}"
and the email_from field would be set automaticaly to the email of that
partner if the email from is not specified.

task-2091859