[IMP] website_form: adapt tests to leftPanel
The tests are now running with the new website_form implementation.

The inputs on the left panel can trigger the target update on "blur"
The text helper does not focus out of the inputs after entering the
text. text_focus_out is introduce here to solve it without rewriting
other tests.

Part of: #42189
task-2092425