[REF] website: introduce snippet versioning scss mixin
The 13.0 snippets (actually the 13.1 ones, still need to find a way to
support the few non compatible changes we made in 13.1) are now
considered the 000 version of our snippet and we are implementing a
system to support old version snippets alongside with new ones.

Commit https://github.com/odoo/odoo/commit/91f5fd0df315287f37229ef5647d544af9c75298
was a preparation of the website folder for this system. Restructuration
of the other app folders will follow.