[FIX] account_asset: Batch 'Generate asset entries' cron
By default, the cron 'Generate asset entries' was executed once a month.
If there were many entries to process, it would just time out and not process
anything, then the same issue would occur everytime the cron restarted.
This cron calls compute_generated_entries, which is also used elsewhere.
Therefore, we modify the cron with a batch_size context in
compute_generated_entries, such that if batch_size is in the context,
we batch the computation and commit every batch.

opw 2118306
opw 2158655