[FIX] survey: use $primary to always get same color in COM and ENT
The goal of using color variable for styling is to get always the same color,
no matter if we are in community or on enterprise.

Before this commit, *$o-brand-primary* was used in survey stylesheet but this
bariable is not corresponding to the same color in COM and in ENT.

After this commit, *$primary* is used to render the same color no matter the
environnement (COM or ENT).

Task ID:
PR #