[FIX] website_mail_channel: get_alias_info don't crash if no access
Before this commit, if you call get_alias_info with an id of group that you cannot read,
it raises a traceback. It was not a problem previously since we don't handle traceback
from rpc call, but now in v13 we toast it.

So instead to check if rpc crash to hide the block, we call the controller with an int
and nor more a slug, and check manually the access_right, access_rule (with search) into
the controller.

How to reproduce the bug:
   Drop a snippet Discussion Group
   Select a channel (public only are shown)
   Save page
   Edit the channel in backend and make it private
   Refresh
     -> a toast with an access error on the mail channel.

After this commit, rpc return empty response, and js code hide the mail group as previously.