[FIX] survey: fix first input autofocus
Before this commit, the survey form widget was focusing on the first text
input found on the page, even if this input was not the first question.

After this commit, the survey form widget focus only on first text input
if that input is the first question on the page.