[FIX] survey: fix survey result filters
Since recent refactoring, filter data have 'question' and 'answers' as
filter_data keys, instead of 'question_text' and 'labels'. This commit fixes
the naming in template using filter_data.

Also, comment that does not count as answer is not linked to a suggested
answer. So that filtering on those comment does nothing because we cannot
find an answer_id in that case.
As it makes no real sens anyway to filter on those kind of comment,
filtering have been disabled on those comments.