[REM] website_event_question: remove unusable questions reporting
PURPOSE

The commit removes the reporting on the event.question model along with its
created SQL view 'event.question.report' because they're simply not usable.

SPECS

Based on latest testing, the reporting was completely unusable if you have more
than a hundred questions (all events combined) and needs to be refactored based
on new specifications.

As we need to re-work event questions to be able to add "free text" questions,
we prefer to temporarily drop the reporting instead of having to adapt it.

LINKS

Upgrade PR odoo/upgrade#799
PR #45497
Task 2189529