[FIX] stock: change of scheduled date
- Install Inventory, Purchase
- General Settings/tick Multi-step routes
- In Warehouses/YourCompany
  Incoming Shipments: Receive goods in input, then quality and then
  stock (3 steps)
- Create a PO and add the same products on 2 lines. One has a scheduled
  date in the future
- Go to transfers, where "Picking List" or "Procurement Group"
  correspond to your PO
  => the 3 scheduled_date correspond to Datetime.now()
- Validate the first picking (WH/IN)

The scheduled_date (resp date_expected) in the remaining pickings WH/INT
(resp. stock moves) are changed to the past

This is due to the fact that both moves are merged in subsequent
pickings. The cleanest solution would be to avoid the merge in case the
expected dates are different. Although this makes sense, that also
implies a change of behavior.

Therefore, we avoid such situation by preventing the propagation of an
expected date if a sibbling move with a smaller date is linked
to the move. In that case the sibbling move is responsible for the
date_expected and not the current move.

opw-2171509