[FIX] core: improve prefetching during onchange
Use case: add a line on an existing a sales order with many lines.
During the onchange on the sales order, the field `product_id.type` is
accessed on all the lines, which are new records with origin for most of
them.  When accessing `product_id.type` on the first line, the field
`product_id` is copied from the first line's origin, then the field
`type` is fetched from the products available from the cache of all
lines, for which only the first one is in cache, so only the first
product is prefetched!  In order to avoid this situation, when copying
`product_id` from the first line's origin, do it for all the lines.