[IMP] mail/im_livechat: move the `anonymous_name` from livechat to mail
Purpose
=======
In livechat, we have an "anonymous name" on the mail channel.
This field is useful because we do not have one partner per visitor.

In social, we also need this field (as we do not have a partner per
Facebook user) so we move it to mail.

Task-2124457
PR #45493
See odoo/enterprise/pull/8481
See odoo/iap-apps/pull/166
See odoo/upgrade/pull/956