[IMP] point_of_sale: Allow to set less than 1 unit in a law context.
In case we're using the set decrease quantity popup, we couldn't set a quantity that was less than 1 without decrease the opposite of the number.
e.g: if i want 0.3 units, i have to click on the product, it will be set to 1. Then i have to decrease the quantity by 0.7.

Now, if the quantity is 1, you are allowed to modify the quantity. Once it's change, it's impossible to decrease anymore.