[IMP] web: Fold column quick create on load examples
Before this commit, after adding a column or loading examples in a
kanban from the quick create column, it stays open. After this commit,
it will stay open if adding a column through it, but it will fold if the
user has loaded a set of columns from an example.

Task ID 2047369
PR #48299