[IMP] mail: add the channels types in the message information
Purpose
=======
Add the channel types in the message information before it's sent
to the frontend. So, we can have different behavior depending on
the channel types.

Be able to not redirect the user when clicking on a name in Discuss,
if this name come from the field `email_from` (as in social, we do
not have the email of the social contacts).

Task-2124457
PR #45493
See odoo/enterprise/pull/8481
See odoo/iap-apps/pull/166
See odoo/upgrade/pull/956