[IMP] crm: allow editing phone and email of the lead
Purpose
=======
We can set a partner on a lead. If we do it, the email and the phone of
the lead are the email/phone of the partner, but we can not edit those
values.

We want to be able to edit them if a partner is set. If we write on the
email/phone of the lead, it should write on the partner (and vice-versa).

Log the changes on the partner.
Track the email and the phone of the partner.

Task-2207636