[FIX] web: Timerange filter issue in list view
Currently, the List view is not allowing to filter the records based on the
Timerange control panel menu as it's only implemented in pivot, graph and
dashboard view.

But the issue reproduce when,
go to the graph view >> filter based on Timerange >> Switch to list view

Now you can see the Timerange filter is working properly in graph view but
once you switch to the list view facet for Timerange will display in the search
panel but it will not consider as the domain in list view records as Timerange
is not implemented for the List view.

Task : 2170634