[FIX] web: tests: detect failed modules in small suites
Before this commit, we waited for 10 seconds before defined the
boot test (checking that all JS modules have been correctly
loaded). However, small test suites (like the mobile one) ended
before this test was defined. So we could end up with a green
branch with tests not being executed (e.g. if a testing module
had missing dependencies, and was part of a small suite).