[FIX] mass_mailing: remove duplicated JS file import
If a user wants to unsubscribe from a mass mailing, he will get a traceback
when doing so. This is due to a file being imported twice. Indeed unsubscribe
file is imported in mass_mailing.layout that calls assets_backend that already
imports unsubscribe.

It seems there is no traceback in 13 but well in 13.2. However we can clean it
in 13. It has been introduced at f296992317e96562c66bd7ad59a5080d6c551ed5 .

Task-2241443

closes odoo/odoo#50722

X-original-commit: 521eb273ce792eb99631fe34faf7d512e3b59a8e
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>