[IMP] point_of_sale: change set quantity behavior in certification context.
Before, when we wanted to decrease quantity of an orderline in a certification context, we had to count for the value we wanted to retrieve.
Now, we allow the user enter the new value he wants and make the calculation to reduce with the correct quantity and creating a new orderline with this calculated value.