[FIX] mail: employee form view flickers on edit and discard
On a large screen (with the chatter displayed on the right of form views),
Go to Hr employee form view >> switch to edit mode  >>  view flickers
As like the chatter disappears for a while and display again. The same issue
occurs when you discard the changes.

There should be no flicker when you edit and discard the changes. the chatter
must always be displayed on the right side of the form.

Flickers occur as the existing chatter widget's element is returned from
renderNode which will remove existing node from node due to which form element
takes whole space and later when _updateView is called chatter will be aded
back which will generate flickering.

To fix this, if chatter instance is available then return fake node which will
be used as hook node, and when whole view is rendered i.e. in _updateView,
replace hook node with chatter.$el.

task-2243856

Co-authored-by: Mohammed Shekha <msh@odoo.com>