[REV] stock: _quantity_done_compute performance
This reverts commit 6802c6af503881bffca064a12a124fe9414db0b3.

Following v13, newId records are instances of NewID and are ignored from
any `read_group` call. This breaks an important feature: when encoding a
stock move line (thus, in an onchange context), the `quantity_done` field
of stock move should be recomputed. We revert the optimization and get
back to a code working in onchange context.

task-2241471