[FIX] mail/website_slides: allow portal to use attachments
Bug
===
When a portal user want to send an attachment on blog/eshop/slide,
he get an 403 error.

In mail.message, during the create operation, we check the read
access right on the attachment. As this access is restricted to
only the internal user, since #38358, an error is raised.

Technical
=========
The verification is done to avoid malicious users to link private
attachments to their messages, to be able to read them.

Task-2214795