[FIX] hr_timesheet: Hide remaining hours tag on task kanban if allow_timesheet is false
Currently, when the project have the 'allow_timesheet' false even though
it will display the remaning hours tag in task kanban view.
so purpose of the task is hide the 'remaining hours' tag if
allow_timesheet is false.

So after this commit, Hide 'remaining_hours' tag on task kanban view
if allow_timesheet is False.

closes odoo/odoo#50754

Taskid: 2243305
Closes: #50357
X-original-commit: daa75ea5f217010deb4cb5937747f37ab533bcd6
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>