[FIX] web: Model base class can dispatch at willStart
Have a component that uses a model with the useModel hook
During the mount, make that model dispatch a command
i.e. in the willStart of the component

Before this commit, there was a timeout because the model waited for
the end of the rendering to yield, and the willStart waited for that
to yield in turn. That was a deadlock

After this commit, we trigger a rendering on the component
only when we are already mounted