[IMP] tests: easier screencasting
Before this commit, screencast frames where only saved through calls to
`wait_code_ok`, which strikes me as weird: once you call
`start_screencast`, you expect it to record all frames sent by the
browser and not just some of them. Move the logic for screencast saving
in the `_get_message` method, since it is used to get all events from
the browser.

Also, add a `stop_screencast` method, and a warning when a screencast
cannot be started properly.