[FIX] base, website: fix post website specific views
Things to know about specific views:

Since the introduction of website specific views, we can have multiple
views that are related to a website (specific views). These views have
no xml_id as it is only set on the original view (generic view).
We rely on the view's key to identify related views as it is the same
for all specific views of a generic view. Also the key is equals to the
xml_id.

When a generic view is updated, we will check if the specific views have
the same values as the generic view. If it is the case, the value that
are the same are considered updatable and are updated on both the
generic and specific view. Else these are considered as noupdate.

When a generic view is created we will check if the potential parent
generic view has specific views. If it is the case we will create a copy
of the created view for each specific view.

The issue:

The method that checks if a created/updated view has specific views is
located in website. When we update a module, each modules are
initialized and updated in a specific order. If a generic view that has
specific views located in a module loaded before website this view's
specific views will not be updated/created as the method does not exist.
This issue is mainly affecting portal views at the moment.

The solution:

For write:
Before website is loaded we can not know if a view is website specific.
We will delay the write of the views that might have a specific
view until website is loaded. Then we will compute all those views
before continuing the update with the same behaviour as before.

As the changes are commited after each module update we can not write
the changes on those before reaching website.
This is due to the noupdate behaviour, if we have a specific view with
values similar to it's generic view and the generic view is updated,
the specific should be updated too. This will not be the case if the
update stops before reaching website as the generic view updated from
the previous update will no longer be similar to the specific one which
has not been updated in the last update due to it being stopped.

For create:
We  will create the generic view as before but wait until we have
updated all the modules to gather all generic views, and create the
specific views.
This will result in a change of behaviour being that if a specific view
that have a parent is deleted it will be recreated on update
(same behaviour as a generic view).
Before the view would not have been recreated.