[IMP] web_editor, website: allow presets user edition
Now, the user can edit each individual component of each color preset
thanks to the third tab of the left panel.

Note: before this commit, the "text" color edition was used to set the
main text color and one of the constrasting color for the color-yiq
function (the darkest if a dark color or the lightest if a light color).
That last part has been removed and automatic colors will now always
use $gray-900 and $white as contrasting colors, unless changed by the
theme explicitely. This needed to be removed as now the "text color"
edition has been moved in the "no-preset" config and should thus only
impact the "no-preset" areas instead of the whole website.

Part of https://github.com/odoo/odoo/pull/45856
task-2197038