[FIX] website_sale: street should never be hidden
Right implementation should have an inverse write on it.

In case of some l10n that split street, in street_name, street_number (l10n_mx)
it is the job of the l10n to hide the street in a l10n_xx_website_sale module
or support at least the inverte write on street and try to parse it.

If a l10n decide to extend ecommerce to had their own field, it is the role of
this module to update code to make street optionnal and hidden if needed.

opw-2241050