[FIX] mrp: add move with different company
Currently the related on the property use the company in the context and
not on the current object. It implies that a raw move created on the MO
form view will have the wrong destination location and trigger the
check company error.

Use a compute instead of related.

fixup! [FIX] mrp: add move with different company

production_location_id not set