[IMP] web: allow Owl Components for widget tag in views
This commit adds the compatibility layer to support both Owl
Components and legacy widgets for the <widget/> tag in basic views
(form, list, kanban). Legacy widgets are still registered in the
widgetRegistry ('web.widget_registry'), but new Owl components must
be registered in the componentRegistry ('web.componentRegistry').

Note that this only concerns widgets/components used when the
<widget/> tag is encountered in one of the 3 main views, and there
are very few usecases for this. However, a new usecase will come
soon in stock, and it would be a shame not to use Owl for it.

Task 2258347