[ADD] odoo-bin cloc cli subcommand.
Odoo cloc is a tool to count the number of relevant lines written in python,
javascript or xml. This can be used as rough metric for pricing maintenance of
customizations.

It has two modes of operation, either by providing a path:

    odoo-bin cloc -p module_path

Or by providing the name of a database:

    odoo-bin cloc --addons-path=dirs -d database

In the later mode, only the custom code is accounted for.