[IMP] account: reconcile models: reduce partner map as much as possible
The partner map becomes a big WITH at the beginning of the query getting the reconciliation candidates from the reconciliation models. It maps each statement line to a partner, even the ones with an explicit partner_id set. These can actually not be kept in the partner map, and gotten using coalesce within the query. We do that here in order to save performances, as on very big databases with a lot of statement lines and unreconciled amls to match, the query takes forever and finally overloads the RAM.

We don't entirely avoid that situation here, but make sure to avoid it as often as possible, and provide a simple solution to circumvent the issue: set partners on the statement lines.