[IMP] stock_picking_batch: add "Put In Pack" button
This commit sets up the logic and adds the button for "Put in Pack"
action that already exists in stock.picking. Unfortunately in order to
handle multiple pickings, a lot of code had to be duplicated for this
logic from stock.picking (i.e. maintanence is probably going to be a
pain). The exception is an extension of the pre-put in pack wizard to
handle products with different destination.

Note that logic to handle "Delivery Packaging" wizard when "Delivery
Methods" was purposely not included in this commit due to complexity of
re-creating wizard due to its stock.picking 'carrier_id' dependency
(i.e. when carrier_id do not match across picking then this option does
not make sense) while also not being a feature used very often.

Part of "1. Improve Batch Pickings" specification of overall barcode
improvements task.

Task: 1884520
Enterprise PR: