[IMP] website(_event): allow to customize placeholder of website search box
PURPOSE

Be more explicit in search boxes: allow to set a placeholder and to display
a "clear filters". It helps users understanding purpose of the search tool.

SPECIFICATIONS

Support giving a placeholder to the website search box. Set it to "Search an
event ..." in website_event.

Allow giving a "clear filters" link to the search tool. If given a small button
is displayed on the right of the box, allowing to clear filters by going back
to a standard page.

LINKS

Task ID 2252655 (master Online Events task)
Task ID 2283796 (website_event B2Basics)