[ADD] website_event_meet: allow the visitors to meet at online events
Purpose
=======
Allow the website visitors to meet at online events using video
conference room.

Specifications
==============
For this purpose, we decided to use Jitsi to create the room, to not
re-create the wheel. So, we can create Jitsi room from website.

Any website visitor can create a room from the frontend.

When a room is full, it is not visible anymore for the website visitor.
This behavior encourage the visitor to create new rooms (and we also
avoid to have to many visitors in a room).

Technical
=========
Jitsi do not provide a REST API (or a similar API) to get the number of
participant in a room (without hosting our own instance of Jitsi).

The only way to get the number of participant is to be in the room and
to use the Javascript API. That's why the participant send in RPC the
number of participant in the room, to update the counter on the server
side.

We can not ask the user who is leaving to update the counter, because he
can close the tab on his browser (and so the event will not be triggered),
so, the first participant in the room send the update in RPC.

Task-2283742