[IMP] website_event_track_online : add wishlist mechanism on event tracks
This commit is part of the event online refactoring.
This adds a wishlist mechanism on event tracks.
Logged in visitors (=res.partner) can add some talks to their wishlist in order
to prepare the event for which they are registrered to.

Some talks can be set as wishlisted by default.
Those talks won't be wishlistable but the visitors but they will appear as
starred on the list of talks. It's not possible to unstar those talks.

An intermediate model have been introduced to manage talks participants:
event.track.partner
This model will be useful not only for the wishlist but also for other type of
participation to the talk, like taking a quiz linked to that talk, etc.
(talk's quiz will be added in later commit)

Task ID: 2284043