[IMP] website_event_online: ensure to link visitor at registration
This commit is part of the event online refactoring.
This ensure that a visitor is created (if it's not yet the case) when a
registration is done on an event.
In case of multiple registrations made by the same visitor, only the first
registration is linked to the created/existing visitor.
We take the opportunity to update the visitor's information based on the
registration's infos.

Task ID: 2284043