[ADD] website_event_track_live: configure Youtube videos / streams on event.tracks
This commit improves the existing event.tracks to include the possibility to
configure a Youtube live stream or video.

The frontend view of the track was adapted accordingly to show the stream /
video instead of the track description when available.

In addition, this commit introduces a "viewers count" feature for relevant
tracks that uses the Youtube Data API to fetch:
- Views count for a regular video or a recorded live stream that is finished
- Currently watching viewers count for a live stream

These API calls require a valid API key that is configurable in the event
settings.
If the call fails (no API key, no quota left, ...), the viewer count
information is simply not displayed.

Task 2283729