[IMP] website_event_online: add countdown on events (momentum)
It is now possible to set a countdown on an event.
On the event main page, if the event hasn't started yet, a countdown is
displayed. At the end of the countdown, there is an automated redirection
to the first talk. If there are more than one talk at the same time,
the redirection is done to the talks list page.

For event manager or event user, this countdown is not shown as they need
to be able to modify the template of introduction page.