[IMP] website_event_track_session: apply display-timer widget on tracks CTA button
This commit applies the display-timer widget to the tracks CTA buttons, if any.
A Pre timer is configured to annonounce the start of the talk, if the visitor
visit the page of the talk a bit to soon.
The main timer depends from the track.website_cta_start_remaining.
At the end of the main timer, the CTA button is displayed to the visitor.