[IMP] websiste_event_track_online: change wishlist widget into reminder widget for tracks
In this commit, the wishlist widget is reworked to use the concept of reminder
instead of wishlist properly speaking.
Users need to be able to un-wishlist key tracks (that are wishlisted by
default). In order to do that, is_blacklisted field have been added on
track_partner model.

When the user un-wishlist a key track via the now called reminder widget,
is_blacklisted is set to True for that track_partner. The user won't receive
reminder anymore for this track.
For normal track, when the user set a reminder, is_wishlisted is set to true,
like it was done before.