[ADD] website_jitsi: allow to create Jitsi room on the frontend
Purpose
=======
Build a mixin to create Jitsi room on website.

Specifications
==============
Add a new model `chat.room` which contains all the configuration of the Jitsi room.

A mixin `chat.room.mixin` add a many2one to this new model, so the behavior can be
optional without adding all fields on the model.

So, we just need to inherit from the mixin and to use the chat room widget to build chat
room on website, for any model.

Technical
=========
Jitsi do not provide a REST API (or a similar API) to get the number of
participant in a room (without hosting our own instance of Jitsi).

The only way to get the number of participant is to be in the room and
to use the Javascript API. That's why the participant send in RPC the
number of participant in the room, to update the counter on the server
side.

We can not ask the user who is leaving to update the counter, because he
can close the tab on his browser (and so the event will not be triggered),
so, the first participant in the room send the update in RPC.

Task-2283742