[ADD] website_event_meet: allow the visitors to meet at online events
Purpose
=======
Allow the website visitors to meet at online events using video
conference room.

Specifications
==============
For this purpose, we decided to use Jitsi to create the room, to not
re-create the wheel. So, we can create Jitsi room from website.

Any website visitor can create a room from the frontend.

When a room is full, it is not visible anymore for the website visitor.
This behavior encourage the visitor to create new rooms (and we also
avoid to have to many visitors in a room).

Task-2283742