[REF] account: rewrite field declaration
In order to have an easier tracking of changes and easier rebases, it
would be nice to have shorter lines in field declaration.
If fields don't fit on one line or if there are too many arguments, they
should follow the pattern:
```python
field = fields.Field(  # no argument on first line
    kw1=value1,        # one and only one argument per line
    kw2=value2,        # all lines end with a comma
)                      # closing parenthesis is on separate line
```

This will avoid useless conflicts, which can be time consuming and error
prone (i.e. two changes on the same line but only one is noticed)
It is also easier to
* find the relevant attributes as they are all at the start of the line
and not anywhere in a big block of code.
* read the diff when reviewing
* check the linter's warnings (removing a lot of E122, E128, E501, E502)