[IMP] stock_picking_batch: implement "cluster picking" for put in pack
This commit adds in "cluster picking" feature which is defined as:
When doing a batch picking, if a product from picking X is assigned to a
destination package then all products in picking X that do not currently
have a destination package assigned will have the same package assigned
to it.

This commit only applies this feature when destination package is
assigned using the "Put in Pack" option. This also supports the feature
at the enterprise level (barcode app).

Task: 1884520
Enterprise PR: