[FIX] website_event_meet: fix "No Open Room" message
Bug
===
If no rooms are opened, a message "No Open Room" will be displayed.

In the case where we are a standard visitor (not event manager) and if
some room are opened, but they are all full, no room will be displayed
but the message "No Open Room" will not be visible.

In that case, we also want to display the message "No Open Room".

(Those room are filtered in QWeb, but if we clean the variable `meeting_rooms`
passed to the template, it will display the message when needed).

Task-2283742