-- waiting for merge --
[IMP] portal, *: move language switcher to portal

*: base, website

Make the language switcher available on portal without website
installed.

task-2203383

Co-authored-by: qsm-odoo <qsm@odoo.com>

[IMP] website, web, portal: new options language selector in footer

Add an option to display the language selection inline or with a
dropdown. Add an option to choose if language labels are flags, text
or both.

task-2203383

Co-authored-by: qsm-odoo <qsm@odoo.com>

[IMP] website: add a field image flag on the lang model

Add a field image flag on the lang model because when a language has
no country (like the arabic) we cannot use the country flag as a lang
flag. By default we use this new field and as fallback we use the
country image. We also add an arabic flag image in the website data.

task-2203383

[IMP] web_editor, *: use images for header/footer

*: website

task-2229551