[FIX] point_of_sale: make _setNumpadMode synchronous
This is a followup of fixing the issues in runbot caused by pos tour tests.

During testing, numpad mode is set and is normally followed by a normal
number input from the numpad widget. Since _setNumpadMode is async, it
could happen that the number input callback is completed first before
setting the numpad mode even if in the tour step, mode is clicked first
before the number input. To fix this, we make the numpad mode callback
to be synchronous. This way, whenever numpad mode is clicked, the callback
is guaranteed to finish first when it is immediately followed by number
click.

And to make sure that the buffer is resetted after changing the mode,
we call NumberBuffer.reset before setting the new mode.